Got the property stuff working
