+++ integration-tests/drivers/TestDriver063/package.json
@@ -26,6 +26,7 @@
     "@babel/core": "^7.10.5",
     "@babel/runtime": "^7.10.5",
     "@react-native-community/eslint-config": "^2.0.0",
+    "@types/detox": "^17.14.3",
     "@types/express": "^4.17.13",
     "assert": "^2.0.0",
     "babel-jest": "^26.1.0",
